requirements.txt
Flask==2.2.5
Flask-SocketIO==5.3.4
requests==2.28.2
python-dotenv==1.0.0
eventlet==0.33.3
